Moved Extension to Microsoft.Toolkit.Uwp, Extracted Regex as separate string.
